Move FlatpakCompletion to app/flatpak-complete.[ch]
