+++ app/flatpak-builtins.h
@@ -24,6 +24,7 @@
 #include <ostree.h>
 #include <gio/gio.h>
 
+#include "flatpak-complete.h"
 #include "flatpak-utils-private.h"
 #include "flatpak-dir-private.h"
